+++ components/main/organisms/footer.tsx
@@ -1,0 +1,7 @@
+export default function Footer(props) {
+  return (
+    <div className="text-center h6 my-4">
+      <div>{props.footer}</div>
+    </div>
+  );
+}
