--- components/organisms/footer.tsx
@@ -1,7 +1,0 @@
-export default function Footer(props) {
-  return (
-    <div className="text-center h6 my-4">
-      <div>{props.footer}</div>
-    </div>
-  );
-}
