--- components/organisms/header.tsx
@@ -1,9 +1,0 @@
-export default function Header(props) {
-  return (
-    <div>
-      <h1 className="bg-primary px-3 text-white display-4 text-right">
-        {props.header}
-      </h1>
-    </div>
-  );
-}
