+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import Layout from '../../components/layout'
+import Layout from '../../components/main/layout'
 
 const IndexPage = () => {
   const [catImageUrl, setCatImageUrl] = useState(
